split material cards into their own files
